refactor(pfb): extract selection state into a useSelection composable
